Convert specification to Bikeshed
This replaces the existing XSLT-based spec-editing pipeline with a more
modern one based on Bikeshed [1]. This was done mostly in an automated
fashion to ensure correctness, using the tooling at [2]. The issues list
at [3] (mostly closed) also contains useful history of the conversion.

This closes #147.

Because of the nature of this change, which touches almost every line,
blame history will be lost. To trace back changes before this commit,
see the old index.xml file at [4].

[1]: https://github.com/tabatkins/bikeshed
[2]: https://github.com/tobie/webidl2bikeshed
[3]: https://github.com/tobie/webidl/issues
[4]: https://github.com/heycam/webidl/blob/dd77b1907e6af7ec27ff67edfb38f892f2fde3a8/index.xml